+++ pymongo/synchronous/encryption.py
@@ -250,15 +250,10 @@ class _EncryptionIO(MongoCryptCallback):  # type: ignore[misc]
         :param database: The database on which to run listCollections.
         :param filter: The filter to pass to listCollections.
 
-        :return: The first document from the listCollections command response as BSON.
+        :return: The all documents from the listCollections command response as BSON.
         """
         with self.client_ref()[database].list_collections(filter=RawBSONDocument(filter)) as cursor:
-            lst = []
-            for doc in cursor:
-                lst.append(_dict_to_bson(doc, False, _DATA_KEY_OPTS))
-            if len(lst) > 0:
-                return lst
-            return None
+            return [_dict_to_bson(doc, False, _DATA_KEY_OPTS) for doc in cursor]
 
     def spawn(self) -> None:
         """Spawn mongocryptd.