+++ test/test_encryption.py
@@ -2410,73 +2410,62 @@ class TestLookupProse(EncryptionIntegrationTest):
     @client_context.require_version_min(7, 0, -1)
     def setUp(self):
         super().setUp()
-        self.encrypted_client = self.rs_or_single_client(
+        encrypted_client = self.rs_or_single_client(
             auto_encryption_opts=AutoEncryptionOpts(
                 key_vault_namespace="db.keyvault",
                 kms_providers={"local": {"key": LOCAL_MASTER_KEY}},
             )
         )
-        self.encrypted_client.db.drop_collection("keyvault")
+        encrypted_client.drop_database("db")
 
         key_doc = json_data("etc", "data", "lookup", "key-doc.json")
-        key_vault = create_key_vault(self.encrypted_client.db.keyvault, key_doc)
+        key_vault = create_key_vault(encrypted_client.db.keyvault, key_doc)
         self.addCleanup(key_vault.drop)
 
-        self.encrypted_client.db.drop_collection("csfle")
-        self.encrypted_client.db.create_collection(
+        encrypted_client.db.create_collection(
             "csfle",
             validator={"$jsonSchema": json_data("etc", "data", "lookup", "schema-csfle.json")},
         )
-
-        self.encrypted_client.db.drop_collection("csfle2")
-        self.encrypted_client.db.create_collection(
+        encrypted_client.db.create_collection(
             "csfle2",
             validator={"$jsonSchema": json_data("etc", "data", "lookup", "schema-csfle2.json")},
         )
-
-        self.encrypted_client.db.drop_collection("qe")
-        self.encrypted_client.db.create_collection(
+        encrypted_client.db.create_collection(
             "qe", encryptedFields=json_data("etc", "data", "lookup", "schema-qe.json")
         )
-
-        self.encrypted_client.db.drop_collection("qe2")
-        self.encrypted_client.db.create_collection(
+        encrypted_client.db.create_collection(
             "qe2", encryptedFields=json_data("etc", "data", "lookup", "schema-qe2.json")
         )
+        encrypted_client.db.create_collection("no_schema")
+        encrypted_client.db.create_collection("no_schema2")
 
-        self.encrypted_client.db.drop_collection("no_schema")
-        self.encrypted_client.db.create_collection("no_schema")
+        unencrypted_client = self.rs_or_single_client()
 
-        self.encrypted_client.db.drop_collection("no_schema2")
-        self.encrypted_client.db.create_collection("no_schema2")
-
-        self.unencrypted_client = self.rs_or_single_client()
-
-        self.encrypted_client.db.csfle.insert_one({"csfle": "csfle"})
-        doc = self.unencrypted_client.db.csfle.find_one()
+        encrypted_client.db.csfle.insert_one({"csfle": "csfle"})
+        doc = unencrypted_client.db.csfle.find_one()
         self.assertTrue(isinstance(doc["csfle"], Binary))
-        self.encrypted_client.db.csfle2.insert_one({"csfle2": "csfle2"})
-        doc = self.unencrypted_client.db.csfle2.find_one()
+        encrypted_client.db.csfle2.insert_one({"csfle2": "csfle2"})
+        doc = unencrypted_client.db.csfle2.find_one()
         self.assertTrue(isinstance(doc["csfle2"], Binary))
-        self.encrypted_client.db.qe.insert_one({"qe": "qe"})
-        doc = self.unencrypted_client.db.qe.find_one()
+        encrypted_client.db.qe.insert_one({"qe": "qe"})
+        doc = unencrypted_client.db.qe.find_one()
         self.assertTrue(isinstance(doc["qe"], Binary))
-        self.encrypted_client.db.qe2.insert_one({"qe2": "qe2"})
-        doc = self.unencrypted_client.db.qe2.find_one()
+        encrypted_client.db.qe2.insert_one({"qe2": "qe2"})
+        doc = unencrypted_client.db.qe2.find_one()
         self.assertTrue(isinstance(doc["qe2"], Binary))
-        self.encrypted_client.db.no_schema.insert_one({"no_schema": "no_schema"})
-        self.encrypted_client.db.no_schema2.insert_one({"no_schema2": "no_schema2"})
+        encrypted_client.db.no_schema.insert_one({"no_schema": "no_schema"})
+        encrypted_client.db.no_schema2.insert_one({"no_schema2": "no_schema2"})
 
     @client_context.require_version_min(8, 1, -1)
-    def test_1(self):
-        self.encrypted_client = self.rs_or_single_client(
+    def test_1_csfle_joins_no_schema(self):
+        encrypted_client = self.rs_or_single_client(
             auto_encryption_opts=AutoEncryptionOpts(
                 key_vault_namespace="db.keyvault",
                 kms_providers={"local": {"key": LOCAL_MASTER_KEY}},
             )
         )
         doc = next(
-            self.encrypted_client.db.csfle.aggregate(
+            encrypted_client.db.csfle.aggregate(
                 [
                     {"$match": {"csfle": "csfle"}},
                     {
@@ -2496,15 +2485,15 @@ class TestLookupProse(EncryptionIntegrationTest):
         self.assertEqual(doc, {"csfle": "csfle", "matched": [{"no_schema": "no_schema"}]})
 
     @client_context.require_version_min(8, 1, -1)
-    def test_2(self):
-        self.encrypted_client = self.rs_or_single_client(
+    def test_2_qe_joins_no_schema(self):
+        encrypted_client = self.rs_or_single_client(
             auto_encryption_opts=AutoEncryptionOpts(
                 key_vault_namespace="db.keyvault",
                 kms_providers={"local": {"key": LOCAL_MASTER_KEY}},
             )
         )
         doc = next(
-            self.encrypted_client.db.qe.aggregate(
+            encrypted_client.db.qe.aggregate(
                 [
                     {"$match": {"qe": "qe"}},
                     {
@@ -2524,15 +2513,15 @@ class TestLookupProse(EncryptionIntegrationTest):
         self.assertEqual(doc, {"qe": "qe", "matched": [{"no_schema": "no_schema"}]})
 
     @client_context.require_version_min(8, 1, -1)
-    def test_3(self):
-        self.encrypted_client = self.rs_or_single_client(
+    def test_3_no_schema_joins_csfle(self):
+        encrypted_client = self.rs_or_single_client(
             auto_encryption_opts=AutoEncryptionOpts(
                 key_vault_namespace="db.keyvault",
                 kms_providers={"local": {"key": LOCAL_MASTER_KEY}},
             )
         )
         doc = next(
-            self.encrypted_client.db.no_schema.aggregate(
+            encrypted_client.db.no_schema.aggregate(
                 [
                     {"$match": {"no_schema": "no_schema"}},
                     {
@@ -2549,15 +2538,15 @@ class TestLookupProse(EncryptionIntegrationTest):
         self.assertEqual(doc, {"no_schema": "no_schema", "matched": [{"csfle": "csfle"}]})
 
     @client_context.require_version_min(8, 1, -1)
-    def test_4(self):
-        self.encrypted_client = self.rs_or_single_client(
+    def test_4_no_schema_joins_qe(self):
+        encrypted_client = self.rs_or_single_client(
             auto_encryption_opts=AutoEncryptionOpts(
                 key_vault_namespace="db.keyvault",
                 kms_providers={"local": {"key": LOCAL_MASTER_KEY}},
             )
         )
         doc = next(
-            self.encrypted_client.db.no_schema.aggregate(
+            encrypted_client.db.no_schema.aggregate(
                 [
                     {"$match": {"no_schema": "no_schema"}},
                     {
@@ -2577,15 +2566,15 @@ class TestLookupProse(EncryptionIntegrationTest):
         self.assertEqual(doc, {"no_schema": "no_schema", "matched": [{"qe": "qe"}]})
 
     @client_context.require_version_min(8, 1, -1)
-    def test_5(self):
-        self.encrypted_client = self.rs_or_single_client(
+    def test_5_csfle_joins_csfle2(self):
+        encrypted_client = self.rs_or_single_client(
             auto_encryption_opts=AutoEncryptionOpts(
                 key_vault_namespace="db.keyvault",
                 kms_providers={"local": {"key": LOCAL_MASTER_KEY}},
             )
         )
         doc = next(
-            self.encrypted_client.db.csfle.aggregate(
+            encrypted_client.db.csfle.aggregate(
                 [
                     {"$match": {"csfle": "csfle"}},
                     {
@@ -2605,15 +2594,15 @@ class TestLookupProse(EncryptionIntegrationTest):
         self.assertEqual(doc, {"csfle": "csfle", "matched": [{"csfle2": "csfle2"}]})
 
     @client_context.require_version_min(8, 1, -1)
-    def test_6(self):
-        self.encrypted_client = self.rs_or_single_client(
+    def test_6_qe_joins_qe2(self):
+        encrypted_client = self.rs_or_single_client(
             auto_encryption_opts=AutoEncryptionOpts(
                 key_vault_namespace="db.keyvault",
                 kms_providers={"local": {"key": LOCAL_MASTER_KEY}},
             )
         )
         doc = next(
-            self.encrypted_client.db.qe.aggregate(
+            encrypted_client.db.qe.aggregate(
                 [
                     {"$match": {"qe": "qe"}},
                     {
@@ -2633,15 +2622,15 @@ class TestLookupProse(EncryptionIntegrationTest):
         self.assertEqual(doc, {"qe": "qe", "matched": [{"qe2": "qe2"}]})
 
     @client_context.require_version_min(8, 1, -1)
-    def test_7(self):
-        self.encrypted_client = self.rs_or_single_client(
+    def test_7_no_schema_joins_no_schema2(self):
+        encrypted_client = self.rs_or_single_client(
             auto_encryption_opts=AutoEncryptionOpts(
                 key_vault_namespace="db.keyvault",
                 kms_providers={"local": {"key": LOCAL_MASTER_KEY}},
             )
         )
         doc = next(
-            self.encrypted_client.db.no_schema.aggregate(
+            encrypted_client.db.no_schema.aggregate(
                 [
                     {"$match": {"no_schema": "no_schema"}},
                     {
@@ -2661,8 +2650,8 @@ class TestLookupProse(EncryptionIntegrationTest):
         self.assertEqual(doc, {"no_schema": "no_schema", "matched": [{"no_schema2": "no_schema2"}]})
 
     @client_context.require_version_min(8, 1, -1)
-    def test_8(self):
-        self.encrypted_client = self.rs_or_single_client(
+    def test_8_csfle_joins_qe(self):
+        encrypted_client = self.rs_or_single_client(
             auto_encryption_opts=AutoEncryptionOpts(
                 key_vault_namespace="db.keyvault",
                 kms_providers={"local": {"key": LOCAL_MASTER_KEY}},
@@ -2670,7 +2659,7 @@ class TestLookupProse(EncryptionIntegrationTest):
         )
         with self.assertRaises(PyMongoError) as exc:
             _ = next(
-                self.encrypted_client.db.csfle.aggregate(
+                encrypted_client.db.csfle.aggregate(
                     [
                         {"$match": {"csfle": "qe"}},
                         {
@@ -2687,8 +2676,8 @@ class TestLookupProse(EncryptionIntegrationTest):
             self.assertTrue("not supported" in str(exc))
 
     @client_context.require_version_max(8, 1, -1)
-    def test_9(self):
-        self.encrypted_client = self.rs_or_single_client(
+    def test_9_error(self):
+        encrypted_client = self.rs_or_single_client(
             auto_encryption_opts=AutoEncryptionOpts(
                 key_vault_namespace="db.keyvault",
                 kms_providers={"local": {"key": LOCAL_MASTER_KEY}},
@@ -2696,7 +2685,7 @@ class TestLookupProse(EncryptionIntegrationTest):
         )
         with self.assertRaises(PyMongoError) as exc:
             _ = next(
-                self.encrypted_client.db.csfle.aggregate(
+                encrypted_client.db.csfle.aggregate(
                     [
                         {"$match": {"csfle": "csfle"}},
                         {