+++ doc/changelog.rst
@@ -8,6 +8,7 @@ PyMongo 4.12 brings a number of changes including:
 
 - Support for configuring DEK cache lifetime via the ``key_expiration_ms`` argument to
   :class:`~pymongo.encryption_options.AutoEncryptionOpts`.
+- Support for $lookup in CSFLE and QE.pr
 
 Issues Resolved
 ...............